+++ NEWS.md
@@ -1,5 +1,10 @@
 # haven 0.2.0.9000
 
+* `as_factor()` gains a new `levels = "default"` mechanism. This uses the
+  labels where present, and otherwise uses the labels. This is now the
+  default, as it seems to map better to the semantics of labelled values
+  in other statistical packages (#81).
+
 * New `vignette("semantics", package = "haven")` discusses the semantics
   of missing values and labelling in SAS, SPSS, and Stata, and how that
   translate